// packages/@mantine-vue/core/src/components/ColorPicker/ColorPicker.types.ts
export type ColorFormat = 'hex' | 'hexa' | 'rgba' | 'rgb' | 'hsl' | 'hsla'
export interface HsvaColor {
  h: number
  s: number
  v: number
  a: number
}
export interface RgbaColor {
  r: number
  g: number
  b: number
  a: number
}
export interface HslaColor {
  h: number
  s: number
  l: number
  a: number
}